requirements.txt
ConfigArgParse==1.2
lmdb==0.98
matplotlib==3.2.1
numpy==1.18.2
python-dateutil==2.8.1
PyYAML==5.3.1
seaborn=0.10.1
scikit-learn==0.22.2.post1
sentencepiece==0.1.90
tensorboard==1.14.0
torch==1.5.0+cu101
torch-cluster==1.5.4
torch-geometric==1.5.0
torch-scatter==2.0.4
torch-sparse==0.6.4
torch-spline-conv==1.2.0
torchfm==0.7.0
torchsearchsorted==1.1
torchtext==0.6.0
torchvision==0.6.0+cu101
tqdm==4.45.0
youtokentome==1.0.6